server-src/planner.c: use g_strdup_printf()
In a lot of places, g_strjoin() was used to build strings where
g_strdup_printf() is more appropriate. Among other things, this allows to remove
four locally declared arrays which sole purposes were to g_sprintf() a number in
it.

What's more, it allows to rename some "lev" local variables to "level", which is
much more comfortable.