server-src/planner.c: some more variable rework in getsize()
First of all, change the name and type of "estimates". This variable is supposed
to host the entire number of estimates to process, so call it "total_estimates"
instead. Also, it was defined as a time_t, which seems a bit weird. Redefine is
as an int.

Finally, introduce a new variable "estimates_for_client": while this role was
devoted to variable "i" prior to this change, the problem is that "i" is too
generic a name for such an important role. Assign it to the value of i in
the appropriate branches and increment total_estimates by this instead of i.

Not only does it makes thing less complicated, it also allows to define
functions aimed at generating estimate request strings and get that logic out of
getsize().