[allocapi] client-src/client_util.c: remove usage of vstrextend()
The only usage was in run_calcsize(), to build a command line out of the
arguments stored in a NULL-terminated GPtrArray, by accessing the structure
members directly. You shouldn't do that.

Do it differently: grab the list of strings from the GPtrArray using
g_ptr_array_free(..., FALSE), and use g_strjoinv() to build the command line
instead. Also, use that list when invoking pipespawnv() instead of accessing
->pdata directly, and when done, use g_strfreev() to free the list.