[allocapi] client-src/runtar.c: remove usage of vstrextend()
The only usage was in main(), to build a command line out of argv. Do it
differently: create a GPtrArray, store quoted arguments of the different quoted
argv arguments, grab the list of strings from the GPtrArray using
g_ptr_array_free(..., FALSE), and use g_strjoinv() to build the command line
instead.