match.c: match_word(): initialize static data pointers earlier
Define a boolean telling whether the separator is NOT '/' (which is probably the
most common case). By default, the "regexes" (a struct mword_regexes *) and
"table" (a struct subst_table *) variables are set to the '/' case, and only set
to the "not '/'" (ie, '.') case if need be.

This further improves the code readability.